+++ backend/src/main/java/com/example/distributed_systems_todoapp/TodoRepository.java
@@ -6,6 +6,6 @@ import jakarta.transaction.Transactional;
 
 
 @Transactional
-public interface TodoRepository extends CrudRepository<Todo, int>{
+public interface TodoRepository extends CrudRepository<Todo, Long>{
     
 }
